Added unit tests, cart implementation (CartController, CartService, CartRepository and CartRepositoryImpl)

+++ src/main/java/com/homework1/onlineshop/domain/Product.java
@@ -4,7 +4,7 @@ import lombok.Builder;
 
 import java.math.BigDecimal;
 @Builder
-record Product(Long id,
+record Product(int id,
                String name,
                BigDecimal price) {
 }
