+++ src/main/java/com/homework1/onlineshop/domain/dto/ProductDto.java
@@ -2,5 +2,5 @@ package com.homework1.onlineshop.domain.dto;
 
 import java.math.BigDecimal;
 
-public record ProductDto(Long id, String name, BigDecimal price) {
+public record ProductDto(int id, String name, BigDecimal price) {
 }
